% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggradient-package.R
\docType{package}
\name{gggradient-package}
\alias{gggradient}
\alias{gggradient-package}
\title{gggradient: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Harry Thompson \email{harry@mayesfield.uk}

}
\keyword{internal}
