% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gggradient-package.R
\docType{package}
\name{gggradient-package}
\alias{gggradient}
\alias{gggradient-package}
\title{gggradient: Scales for Gradient Fills in 'ggplot2'}
\description{
Scales to map continuous x or y data values onto the fill aesthetic, producing a gradient fill along an axis in a 'ggplot2' plot. Set continuous or binned scales that vary fill across individual shapes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hrryt/gggradient}
  \item \url{https://hrryt.github.io/gggradient/}
  \item Report bugs at \url{https://github.com/hrryt/gggradient/issues}
}

}
\author{
\strong{Maintainer}: Harry Thompson \email{harry@mayesfield.uk}

}
\keyword{internal}
